pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tapestry"
version = "0.1.0"
description = "AI-native web intelligence workflow for crawling, organizing, and synthesizing web content"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Tapestry Contributors"}
]
keywords = ["web-scraping", "knowledge-base", "ai", "content-extraction", "crawler"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Core dependencies required for basic functionality
dependencies = [
    "httpx>=0.27.0",           # Async HTTP client for fetching
    "Markdown>=3.6",           # Full Markdown rendering for the static viewer
    "pydantic>=2.0.0",         # Data validation and models
    "pymdown-extensions>=10.0", # Task lists and fenced-code extensions for Markdown
    "selectolax>=0.3.0",       # Fast HTML parsing with CSS selectors
    "readability-lxml>=0.8.0", # Content extraction fallback
    "chardet>=5.0.0",          # Character encoding detection for readability
]

[project.optional-dependencies]
# Browser rendering support for JavaScript-heavy sites
browser = [
    "playwright>=1.40.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# All optional features
all = [
    "tapestry[browser,dev]",
]

[project.urls]
Homepage = "https://github.com/NatsuFox/Tapestry"
Repository = "https://github.com/NatsuFox/Tapestry"
Issues = "https://github.com/NatsuFox/Tapestry/issues"

[tool.setuptools]
packages = ["_src", "_src.crawlers"]

[tool.setuptools.package-dir]
"" = "skills/tapestry"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=skills/tapestry/_src --cov-report=term-missing"

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "selectolax.*",
    "readability.*",
    "playwright.*",
]
ignore_missing_imports = true
